refactor(create-team-page): rewrite team creation form component
Remove old form.js and replace with a new TypeScript version that uses
escape-html for safer team name rendering. Add typings for Ember services
and model to improve type safety and maintainability. This change improves
code clarity, type checking, and security by escaping HTML properly.
