+++ package.json
@@ -75,6 +75,14 @@
     "@types/culori": "^4.0.1",
     "@types/dompurify": "^3.2.0",
     "@types/ember": "^4.0.11",
+    "@types/ember-cli-fastboot": "^2.2.6",
+    "@types/ember-data": "^4.4.16",
+    "@types/ember-data__adapter": "^4.0.6",
+    "@types/ember-data__model": "^4.0.5",
+    "@types/ember-data__serializer": "^4.0.6",
+    "@types/ember-data__store": "^4.0.7",
+    "@types/ember-qunit": "^6.1.3",
+    "@types/ember-resolver": "^9.0.2",
     "@types/ember__application": "^4.0.11",
     "@types/ember__array": "^4.0.10",
     "@types/ember__component": "^4.0.22",
@@ -93,14 +101,7 @@
     "@types/ember__test": "^4.0.6",
     "@types/ember__test-helpers": "^2.9.3",
     "@types/ember__utils": "^4.0.7",
-    "@types/ember-cli-fastboot": "^2.2.6",
-    "@types/ember-data": "^4.4.16",
-    "@types/ember-data__adapter": "^4.0.6",
-    "@types/ember-data__model": "^4.0.5",
-    "@types/ember-data__serializer": "^4.0.6",
-    "@types/ember-data__store": "^4.0.7",
-    "@types/ember-qunit": "^6.1.3",
-    "@types/ember-resolver": "^9.0.2",
+    "@types/escape-html": "^1.0.4",
     "@types/mustache": "^4.2.6",
     "@types/prismjs": "^1.26.5",
     "@types/qunit": "^2.19.13",
@@ -215,6 +216,7 @@
     "ansi_up": "^6.0.6",
     "array-to-sentence": "^2.0.0",
     "dompurify": "^3.3.1",
+    "escape-html": "^1.0.3",
     "fuse.js": "^7.0.0",
     "lottie-web": "^5.13.0",
     "mustache": "^4.2.0",
